refactor(agent): split validate-sql logic from wiring for honest coverage
Per maintainer guidance on the codecov gap: cover what's unit-testable
without an architectural change, and ignore only the genuinely
integration-only code.

- New validate_sql_check.go holds the pure logic (batch loop
  validateStatements, the size cap checkSQLBatchSize, and cap-env parsing
  sqlValidationMaxStatements) — unit-tested and NOT codecov-ignored, so
  regressions in it are caught.
- validate_sql.go now contains only the Mongo/warehouse wiring (claim,
  load, write, run entrypoints). Unit-testing it would require a real
  Mongo/warehouse (integration — not run in the unit coverage job) or a
  DI/interface refactor (architectural), so it's added back to the codecov
  ignore list alongside index_schema.go and the *_repo.go files, all of
  which follow the same integration-tested-and-ignored convention.

This addresses the Copilot concern (no unit-tested logic hidden behind the
ignore) while clearing the informational codecov/patch gap. No behaviour
change; unit + integration tests unchanged and green.
