+++ codecov.yml
@@ -9,6 +9,7 @@ ignore:
   - "services/agent/main.go"
   - "services/agent/agentserver/agentserver.go" # startup/wiring — tested by integration tests
   - "services/agent/agentserver/index_schema.go" # startup/wiring (--mode index-schema) — covered by integration tests
+  - "services/agent/agentserver/validate_sql.go" # Mongo/warehouse wiring (--mode validate-sql) — covered by integration tests; pure logic is in validate_sql_check.go (not ignored)
   - "services/api/internal/server/server.go"     # HTTP routing/wiring — covered by integration tests
   - "services/api/internal/backfill/"           # CLI tool — tested manually
   - "services/api/database/*_repo.go"           # MongoDB repos — covered by integration tests
